move raw_input function to helper module
